refactor: rename Todo components to Note and update copyright headers
- Rename TodoDetail to NoteDetail screen and component
- Rename TodoList to NoteList screen and component
- Rename TodoItem to NoteItem component
- Rename TodoActionButton to NoteActionButton component
- Rename TodoUpdateButton to NoteUpdateButton component
- Update all copyright headers from "React Native Todo App" to "Carrot Note App"
- Update component imports and references throughout navigation and screens
