+++ app/components/noteActionButton/index.ts
@@ -3,11 +3,4 @@
  This software is free to use, modify, and share under 
  the terms of the GNU General Public License v3.
 */
-export interface Todo {
-  id: string;
-  todo: string;
-  completed: boolean;
-  createdAt: string;
-  category?: string;
-  userId?: string;
-}
+export { NoteActionButton } from './noteActionButton';
