--- app/components/todoUpdateButton/index.ts
@@ -1,6 +1,0 @@
-/*
- Copyright (C) 2025 Michael Lim - React Native Todo App 
- This software is free to use, modify, and share under 
- the terms of the GNU General Public License v3.
-*/
-export { TodoUpdateButton } from './todoUpdateButton';
